refactor tests and add one new snapshot
